refactor: split System into one trait per OS resource
system.rs, os_system.rs and system_mock.rs had grown too big to read
comfortably.

System is now split into FileSystem, Process, Network, Terminal and
Host, and stays on as a supertrait over them so callers keep working
unchanged. Both implementations follow the same split, so everything
about one resource sits in one file.

Each mock is named after its resource and ends in Mock, and its builder
methods sit next to the trait they configure.

The mock also drops tests that only proved a value could be stored and
read back, along with a builder that nothing ever called.
