+++ src/platform.rs
@@ -1,12 +1,39 @@
+mod file_system;
+mod host;
+mod network;
+mod os_file_system;
+mod os_host;
+mod os_network;
+mod os_process;
 mod os_system;
+mod os_terminal;
+mod process;
 mod read_write;
 mod stream;
 mod system;
-mod system_mock;
+mod terminal;
 
+#[cfg(test)]
+mod file_system_mock;
+#[cfg(test)]
+mod host_mock;
+#[cfg(test)]
+mod network_mock;
+#[cfg(test)]
+mod process_mock;
+#[cfg(test)]
+mod system_mock;
+#[cfg(test)]
+mod terminal_mock;
+
+pub use file_system::*;
+pub use host::*;
+pub use network::*;
 pub use os_system::*;
+pub use process::*;
 pub use read_write::*;
 pub use stream::*;
 pub use system::*;
 #[cfg(test)]
-pub use system_mock::tests::SystemMock;
+pub use system_mock::SystemMock;
+pub use terminal::*;
